tests/requirements.txt
httpx
jsonschema
pytest
pytest-cov
pyyaml
